Refactor LoanOfferServiceImpl class: divide methods and move calculation logic to CreditParamsCalculator class
